Refactored Folder structure 3rd party sign-in flow returns session token in redirect url
