+++ Controllers/AuthController.cs
@@ -1,4 +1,4 @@
-using MediaTrackerAuthenticationService.Models;
+using MediaTrackerAuthenticationService.Models.Utils;
 using Microsoft.AspNetCore.Mvc;
 using System.Net.Http;
 using MediaTrackerAuthenticationService.Services.AuthService;
@@ -21,14 +21,21 @@ public class Auth : ControllerBase
     public ActionResult<ServiceResponse<string>> Get()
     {
         var response = _authService.GetGoogle();
-        Response.Cookies.Append("SessionToken", "sdfdsfds", new CookieOptions
-        {
-            IsEssential=true,
-            Secure=false,
-            Domain="http://127.0.0.1",
-            SameSite = SameSiteMode.None
-            
-        });
+        // Response.Cookies.Append("SessionToken", "sdfdsfds", new CookieOptions
+        // {
+        //     IsEssential=true,
+        //     Path="/"
+        // });
+
+
+        // // // Response.Headers.Add("Access-Control-Allow-Origin", "http://localhost:5173");
+        // // // Response.Headers.Add("Test", "test");
+        // // // return Ok();
+        Response.Headers.Add("Access-Control-Allow-Origin", "*");
+        // Response.Headers.Add("Access-Control-Allow-Credentials", "true");
+        
+        // // Response.Cookies.Append("MyCookie", "CookieValue");
+        // return Redirect("https://www.google.com/");
         return Ok(response);
     }
 
@@ -48,13 +55,6 @@ public class Auth : ControllerBase
         {
             return StatusCode(500, response); //TODO: change to correct status code and maybe add a better message
         }
-
-        Response.Cookies.Append("SessionToken", "sdfdsfds", new CookieOptions
-        {
-            HttpOnly = true, // The cookie is only accessible via HTTP requests
-            Secure = true, // Ensure that the cookie is only sent over HTTPS
-            SameSite = SameSiteMode.None, // Adjust this based on your application's requirements
-        });
 
         return Redirect(response.Data);
     }
