+++ Services/RequestUrlBuilderService/IRequestUrlBuilderService.cs
@@ -1,4 +1,4 @@
-using MediaTrackerAuthenticationService.Models;
+using MediaTrackerAuthenticationService.Models.Utils;
 
 namespace MediaTrackerAuthenticationService.Services.RequestUrlBuilderService
 {
